modify distillation script to add sft instead
